requirements.txt
langchain
langchain-community
langchain-core
openai
pypdf
numpy
pandas
tiktoken
aiofiles
fastapi[all]
uvicorn
jinja2
python-multipart
pyPDF2
faiss-cpu
python-dotenv
-e .